Addressing review feedback
