+++ modules/loader/src/main/scala/com/snowplowanalytics/snowplow/rdbloader/algebras/db/MigrationBuilder.scala
@@ -10,7 +10,7 @@
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
  */
-package com.snowplowanalytics.snowplow.rdbloader.algerbas.db
+package com.snowplowanalytics.snowplow.rdbloader.algebras.db
 
 import cats.{Applicative, Monad, ~>}
 import cats.data.EitherT
