+++ modules/redshift-loader/src/main/scala/com/snowplowanalytics/snowplow/loader/redshift/db/RedshiftManifest.scala
@@ -15,7 +15,7 @@ import cats.Monad
 import cats.data.NonEmptyList
 import cats.implicits._
 import com.snowplowanalytics.iglu.schemaddl.redshift._
-import com.snowplowanalytics.snowplow.rdbloader.algerbas.db.Manifest
+import com.snowplowanalytics.snowplow.rdbloader.algebras.db.Manifest
 import com.snowplowanalytics.snowplow.rdbloader.common.LoaderMessage
 import com.snowplowanalytics.snowplow.rdbloader.common.S3.Folder
 import com.snowplowanalytics.snowplow.rdbloader.dsl.Logging
