+++ modules/snowflake-loader/src/main/scala/com/snowplowanalytics/snowplow/loader/snowflake/db/RedshiftManifest.scala
@@ -5,7 +5,7 @@ import cats.data.NonEmptyList
 import cats.effect.MonadThrow
 import cats.implicits._
 import com.snowplowanalytics.iglu.schemaddl.redshift._
-import com.snowplowanalytics.snowplow.rdbloader.algerbas.db.{Manifest, Transaction}
+import com.snowplowanalytics.snowplow.rdbloader.algebras.db.{Manifest, Transaction}
 import com.snowplowanalytics.snowplow.rdbloader.common.LoaderMessage
 import com.snowplowanalytics.snowplow.rdbloader.common.S3.Folder
 import com.snowplowanalytics.snowplow.rdbloader.db.Transaction
